requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.17.0
altgraph==0.17.4
asyncpg==0.30.0
attrs==25.1.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
diffusers==0.33.1
disnake==2.10.1
docx==0.2.4
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.2
greenlet==3.1.1
huggingface-hub==0.31.1
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
lxml==5.3.1
Mako==1.3.10
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.1.0
propcache==0.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.11
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.1
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
schedule==1.2.2
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.39
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
yarl==1.18.3
zipp==3.21.0
